vulkan, av1: Remove erroneous driver bug check
AV1 decoders do not have to advertise support for output distinct to
handle the filmgrain case.

"When <<decode-av1-film-grain,film grain is enabled>> for the decoded frame,
the application is required to specify a different decode output picture
resource in slink:VkVideoDecodeInfoKHR::pname:dstPictureResource compared
to the reconstructed picture specified in
slink:VkVideoDecodeInfoKHR::pname:pSetupReferenceSlot->pPictureResource
even if the implementation does not report support for
ename:VK_VIDEO_DECODE_CAPABILITY_DPB_AND_OUTPUT_DISTINCT_BIT_KHR in
slink:VkVideoDecodeCapabilitiesKHR::pname:flags for the
video decode profile."

NVIDIA doesn't do this for example, which is how this bug was caught.